validate_setup.py
#!/usr/bin/env python3
"""
AI Provider Setup Validation Script

This script validates your AI provider configuration and tests the connection
to ensure everything is working correctly.
"""

import os
import sys
from pathlib import Path
from typing import Dict, Any, Optional

# Try to load environment variables
try:
    from dotenv import load_dotenv
    load_dotenv()
except ImportError:
    print("💡 Install python-dotenv for .env file support: pip install python-dotenv")

def check_file_structure():
    """Check if all required files are present."""
    print("📁 Checking file structure...")
    
    required_files = [
        "tools/__init__.py",
        "tools/registry.py",
        "tools/github_tools.py",
        "tools/analysis_tools.py",
        "tools/ai_analysis_tools.py",
        "tools/filesystem_tools.py",
        "tools/communication_tools.py",
        "state.py",
        "nodes.py",
        "workflow.py",
        "api.py",
        "requirements.txt",
        "test_runner.py",
        "TESTING.md"
    ]
    
    missing_files = []
    for file_path in required_files:
        if not Path(file_path).exists():
            missing_files.append(file_path)
        else:
            print(f"   ✓ {file_path}")
    
    if missing_files:
        print(f"\n❌ Missing files:")
        for file_path in missing_files:
            print(f"   - {file_path}")
        return False
    
    print(f"   ✅ All {len(required_files)} required files present")
    return True


def check_python_version():
    """Check Python version compatibility."""
    print("\n🐍 Checking Python version...")
    
    version = sys.version_info
    print(f"   Python version: {version.major}.{version.minor}.{version.micro}")
    
    if version.major < 3 or (version.major == 3 and version.minor < 8):
        print("   ❌ Python 3.8+ required")
        return False
    
    print("   ✅ Python version compatible")
    return True


def check_dependencies():
    """Check if dependencies can be imported."""
    print("\n📦 Checking dependencies...")
    
    dependencies = [
        ("pydantic", "Core data validation"),
        ("langchain", "LangChain framework"),
        ("langgraph", "LangGraph workflow"),
        ("fastapi", "API framework"),
        ("aiohttp", "Async HTTP client"),
    ]
    
    available = []
    missing = []
    
    for dep_name, description in dependencies:
        try:
            __import__(dep_name)
            available.append((dep_name, description))
            print(f"   ✓ {dep_name} - {description}")
        except ImportError:
            missing.append((dep_name, description))
            print(f"   ❌ {dep_name} - {description}")
    
    if missing:
        print(f"\n⚠️  Missing dependencies ({len(missing)}):")
        print("   Install with: pip install -r requirements.txt")
        for dep_name, description in missing:
            print(f"   - {dep_name}")
    
    return len(missing) == 0


def check_environment_variables():
    """Check environment variables."""
    print("\n🔑 Checking environment variables...")
    
    env_vars = [
        ("GITHUB_TOKEN", "GitHub API access", False),
        ("XAI_API_KEY", "Grok (X.AI) API access", False),
        ("SLACK_WEBHOOK_URL", "Slack notifications", True),
        ("EMAIL_USERNAME", "Email notifications", True),
        ("JIRA_URL", "Jira integration", True),
    ]
    
    configured = []
    missing = []
    
    for var_name, description, optional in env_vars:
        value = os.getenv(var_name)
        if value:
            configured.append((var_name, description))
            # Don't print the actual value for security
            print(f"   ✓ {var_name} - {description} (configured)")
        else:
            missing.append((var_name, description, optional))
            status = "optional" if optional else "recommended"
            print(f"   ⚠️  {var_name} - {description} ({status})")
    
    if missing:
        print(f"\n💡 Environment setup tips:")
        print("   1. Copy .env.example to .env")
        print("   2. Add your API keys to .env file")
        print("   3. Source the .env file or restart your terminal")
        
        required_missing = [item for item in missing if not item[2]]
        if required_missing:
            print(f"\n⚠️  For full functionality, configure:")
            for var_name, description, _ in required_missing:
                print(f"   - {var_name}: {description}")
    
    return len([item for item in missing if not item[2]]) == 0


def check_tool_files():
    """Check tool files for basic syntax."""
    print("\n🔧 Checking tool files...")
    
    tool_files = [
        "tools/registry.py",
        "tools/github_tools.py", 
        "tools/analysis_tools.py",
        "tools/ai_analysis_tools.py",
        "tools/filesystem_tools.py",
        "tools/communication_tools.py"
    ]
    
    syntax_ok = True
    
    for file_path in tool_files:
        try:
            with open(file_path, 'r') as f:
                content = f.read()
            
            # Basic syntax check
            compile(content, file_path, 'exec')
            print(f"   ✓ {file_path} - syntax OK")
            
        except SyntaxError as e:
            print(f"   ❌ {file_path} - syntax error: {e}")
            syntax_ok = False
        except Exception as e:
            print(f"   ⚠️  {file_path} - {e}")
    
    return syntax_ok


def get_setup_instructions():
    """Get setup instructions based on current state."""
    instructions = []
    
    # Check if dependencies are missing
    try:
        import pydantic
        import langchain
    except ImportError:
        instructions.append("1. Install dependencies: pip install -r requirements.txt")
    
    # Check environment variables
    if not os.getenv("GITHUB_TOKEN"):
        instructions.append("2. Get a GitHub personal access token from https://github.com/settings/tokens")
        instructions.append("   - Add it to .env file as GITHUB_TOKEN=your_token_here")
    
    if not os.getenv("XAI_API_KEY"):
        instructions.append("3. Get a Grok (X.AI) API key from https://console.x.ai/")
        instructions.append("   - Add it to .env file as XAI_API_KEY=your_key_here")
    
    if not instructions:
        instructions.append("✅ Setup looks good! You can now run: python test_runner.py")
    
    return instructions


def test_ai_provider():
    """Test the configured AI provider."""
    print("\n🤖 Testing AI provider...")

    try:
        from tools.ai_analysis_tools import get_ai_config, GenericAILLM

        # Get AI configuration
        config = get_ai_config()
        print(f"   Provider: {config.provider.value}")
        print(f"   Model: {config.model_name}")

        # Test connection with a simple prompt
        llm = GenericAILLM(config)
        test_prompt = "Hello! Please respond with 'AI provider working correctly.'"
        response = llm.invoke([{"role": "user", "content": test_prompt}])

        if response and len(response) > 0:
            print(f"   ✅ AI Provider Test: SUCCESS")
            print(f"   📝 Response: {response[:50]}...")
            return True
        else:
            print("   ❌ AI Provider Test: FAILED - Empty response")
            return False

    except ImportError as e:
        print(f"   ⚠️  Cannot test AI provider - missing dependencies: {e}")
        print("   💡 Install requirements: pip install -r requirements.txt")
        return True  # Don't fail validation for missing optional deps
    except Exception as e:
        print(f"   ❌ AI Provider Test: FAILED - {e}")
        print("   💡 Check your API key and provider configuration")
        return False

def main():
    """Main validation function."""
    print("🚀 CustomLangGraphChatBot Setup Validation")
    print("=" * 50)
    
    checks = [
        ("File Structure", check_file_structure),
        ("Python Version", check_python_version),
        ("Dependencies", check_dependencies),
        ("Environment Variables", check_environment_variables),
        ("Tool Files", check_tool_files),
        ("AI Provider", test_ai_provider),
    ]
    
    results = {}
    for check_name, check_func in checks:
        try:
            results[check_name] = check_func()
        except Exception as e:
            print(f"   ❌ {check_name} check failed: {e}")
            results[check_name] = False
    
    # Summary
    print("\n" + "=" * 50)
    print("📊 Validation Summary:")
    
    passed = sum(results.values())
    total = len(results)
    
    for check_name, result in results.items():
        status = "✅ PASS" if result else "❌ FAIL"
        print(f"   {status} {check_name}")
    
    print(f"\nOverall: {passed}/{total} checks passed")
    
    # Setup instructions
    if passed < total:
        print(f"\n🔧 Next Steps:")
        instructions = get_setup_instructions()
        for instruction in instructions:
            print(f"   {instruction}")
    else:
        print(f"\n🎉 All checks passed! Your setup is ready.")
        print(f"   Run 'python test_runner.py' to test the tools system.")

    if passed < total:
        print(f"\n💡 For AI provider setup help, see: AI_PROVIDERS_SETUP.md")
    
    return passed == total


if __name__ == "__main__":
    success = main()
    sys.exit(0 if success else 1)
